Fix Live Message Pruning
- Live messages don't get pruned as they get resent in the chat
- Added new field in the database table for live status indication
- Add live messages back into tracker and only edit it from then on instead of it being resent after every check

+++ src/bot/utility/eventSender.js
@@ -1,5 +1,6 @@
 const {SlashCommandBuilder, ChannelType, EmbedBuilder, ButtonStyle, ButtonBuilder, ActionRowBuilder} = require('discord.js');
 const {DateTime, Settings} = require('luxon');
+const {insertUpdateMessage} = require('../../scraper/sqliteHelper')
 
 function embedBuilder(queryResults, getLive = false) {
     let startString = DateTime.fromISO(queryResults["start_time"]).setZone("America/New_York").toLocaleString({ weekday: 'short', month: 'short', day: '2-digit', hour: '2-digit', minute: '2-digit' });
@@ -97,9 +98,11 @@ function updateMessage(queryResults, channelID, messageID) {
             let embed = embedBuilder(queryResults,true);
             let live = embed['live'];
             delete embed['live'];
-            if(live) {
+            if(live && queryResults["liveStatus"] == 0) {
                 message.delete();
-                channel.send(embed);
+                channel.send(embed).then(message => {
+                    insertUpdateMessage(message.id, channel.id, queryResults["eventId"],JSON.stringify(queryResults),queryResults["end_time"], 1);
+                })
             } else {
                 message.edit(embed);
             }
